Move HS and rayleigh sponge into src
