+++ examples/hybrid/sphere/baroclinic_wave_utilities.jl
@@ -12,102 +12,6 @@ include("../staggered_nonhydrostatic_model.jl")
 include("./topography.jl")
 include("../initial_conditions.jl")
 
-##
-## Additional tendencies
-##
-
-# Rayleigh sponge
-
-function rayleigh_sponge_cache(
-    Y,
-    dt;
-    zd_rayleigh = FT(15e3),
-    α_rayleigh_uₕ = FT(1e-4),
-    α_rayleigh_w = FT(1),
-)
-    ᶜz = Fields.coordinate_field(Y.c).z
-    ᶠz = Fields.coordinate_field(Y.f).z
-    ᶜαₘ_uₕ = @. ifelse(ᶜz > zd_rayleigh, α_rayleigh_uₕ, FT(0))
-    ᶠαₘ_w = @. ifelse(ᶠz > zd_rayleigh, α_rayleigh_w, FT(0))
-    zmax = maximum(ᶠz)
-    ᶜβ_rayleigh_uₕ =
-        @. ᶜαₘ_uₕ * sin(FT(π) / 2 * (ᶜz - zd_rayleigh) / (zmax - zd_rayleigh))^2
-    ᶠβ_rayleigh_w =
-        @. ᶠαₘ_w * sin(FT(π) / 2 * (ᶠz - zd_rayleigh) / (zmax - zd_rayleigh))^2
-    return (; ᶜβ_rayleigh_uₕ, ᶠβ_rayleigh_w)
-end
-
-function rayleigh_sponge_tendency!(Yₜ, Y, p, t, colidx)
-    (; ᶜβ_rayleigh_uₕ) = p
-    @. Yₜ.c.uₕ[colidx] -= ᶜβ_rayleigh_uₕ[colidx] * Y.c.uₕ[colidx]
-end
-
-forcing_cache(Y, ::Nothing) = NamedTuple()
-
-# Held-Suarez forcing
-
-forcing_cache(Y, ::HeldSuarezForcing) = (;
-    ᶜσ = similar(Y.c, FT),
-    ᶜheight_factor = similar(Y.c, FT),
-    ᶜΔρT = similar(Y.c, FT),
-    ᶜφ = deg2rad.(Fields.coordinate_field(Y.c).lat),
-)
-
-function held_suarez_tendency!(Yₜ, Y, p, t, colidx)
-    (; T_sfc, z_sfc, ᶜp, ᶜσ, ᶜheight_factor, ᶜΔρT, ᶜφ, params) = p # assume ᶜp has been updated
-
-    R_d = FT(CAP.R_d(params))
-    κ_d = FT(CAP.kappa_d(params))
-    cv_d = FT(CAP.cv_d(params))
-    day = FT(CAP.day(params))
-    MSLP = FT(CAP.MSLP(params))
-    grav = FT(CAP.grav(params))
-
-    z_bottom = Spaces.level(Fields.coordinate_field(Y.c).z[colidx], 1)
-    z_surface = Fields.Field(Fields.field_values(z_sfc[colidx]), axes(z_bottom))
-
-    σ_b = FT(7 / 10)
-    k_a = 1 / (40 * day)
-    k_s = 1 / (4 * day)
-    k_f = 1 / day
-    if :ρq_tot in propertynames(Y.c)
-        ΔT_y = FT(65)
-        T_equator = FT(294)
-    else
-        ΔT_y = FT(60)
-        T_equator = FT(315)
-    end
-    Δθ_z = FT(10)
-    T_min = FT(200)
-
-    @. ᶜσ[colidx] =
-        ᶜp[colidx] / (MSLP * exp(-grav * z_surface / R_d / T_sfc[colidx]))
-
-    @. ᶜheight_factor[colidx] = max(0, (ᶜσ[colidx] - σ_b) / (1 - σ_b))
-    @. ᶜΔρT[colidx] =
-        (k_a + (k_s - k_a) * ᶜheight_factor[colidx] * (cos(ᶜφ[colidx])^2)^2) *
-        Y.c.ρ[colidx] *
-        ( # ᶜT - ᶜT_equil
-            ᶜp[colidx] / (Y.c.ρ[colidx] * R_d) - max(
-                T_min,
-                (
-                    T_equator - ΔT_y * sin(ᶜφ[colidx])^2 -
-                    Δθ_z * log(ᶜp[colidx] / MSLP) * cos(ᶜφ[colidx])^2
-                ) * fast_pow(ᶜσ[colidx], κ_d),
-            )
-        )
-
-    @. Yₜ.c.uₕ[colidx] -= (k_f * ᶜheight_factor[colidx]) * Y.c.uₕ[colidx]
-    if :ρθ in propertynames(Y.c)
-        @. Yₜ.c.ρθ[colidx] -= ᶜΔρT[colidx] * fast_pow((MSLP / ᶜp[colidx]), κ_d)
-    elseif :ρe_tot in propertynames(Y.c)
-        @. Yₜ.c.ρe_tot[colidx] -= ᶜΔρT[colidx] * cv_d
-    elseif :ρe_int in propertynames(Y.c)
-        @. Yₜ.c.ρe_int[colidx] -= ᶜΔρT[colidx] * cv_d
-    end
-    return nothing
-end
-
 # 0-Moment Microphysics
 
 microphysics_cache(Y, ::Nothing) = NamedTuple()
