+++ examples/hybrid/utilities.jl
@@ -81,7 +81,3 @@ function verify_callbacks(t)
         )
     end
 end
-
-# we may be hitting a slow path:
-# https://stackoverflow.com/questions/14687665/very-slow-stdpow-for-bases-very-close-to-1
-fast_pow(x::FT, y::FT) where {FT <: AbstractFloat} = exp(y * log(x))
